Use script to instead dockerfile steps

+++ build.sh
@@ -15,30 +15,9 @@ else
     echo "======= $CONTAINER_NAME_TOOLCHAIN will run."
 fi
 
-wget -O jdk.rpm "$JDK_RPM_SOURCE"
-
-if [[ ! -f "jdk.rpm" ]]; then
-    echo "jdk.rpm found"
-    exit 1
-fi
-
-cd sr-toolchain
-rm -rf starrocks
-git clone -b $GIT_BRANCH $GIT_REPO
-
-if [[ ! -d "starrocks" ]]; then  
-    echo "starrocks not found"
-    exit 1
-fi 
-
-if [[ ! -f "starrocks/thirdparty/vars.sh" ]]; then
-    echo "vars.sh not found"
-    exit 1
-fi
-
-cp ../jdk.rpm .
-if [[ ! -f "jdk.rpm" ]]; then
-    echo "jdk not found"
+wget -O java.tar.gz "$JDK_SOURCE"
+if [[ ! -f "java.tar.gz" ]]; then
+    echo "java.tar.gz found"
     exit 1
 fi
 
@@ -46,28 +25,48 @@ wget -O cmake.tar "$CMAKE_SOURCE"
 rm -rf cmake && mkdir cmake && tar -xvf cmake.tar -C cmake --strip-components 1
 rm -rf cmake.tar
 
-copy_num=$(sed -n  '/===== Downloading thirdparty archives...done/=' starrocks/thirdparty/download-thirdparty.sh)
+
+# build toolchain
+echo "========== start to build $IMAGE_NAME_TOOLCHAIN..."
+rm -rf sr-toolchain/starrocks
+
+git clone -b $GIT_BRANCH $GIT_REPO sr-toolchain/starrocks
+
+if [[ ! -d "sr-toolchain/starrocks" ]]; then  
+    echo "starrocks not found"
+    exit 1
+fi 
+
+if [[ ! -f "sr-toolchain/starrocks/thirdparty/vars.sh" ]]; then
+    echo "vars.sh not found"
+    exit 1
+fi
+
+cp java.tar.gz sr-toolchain/
+if [[ ! -f "sr-toolchain/java.tar.gz" ]]; then
+    echo "jdk not found"
+    exit 1
+fi
+
+cp -r cmake sr-toolchain/
+cp install_env_gcc.sh sr-toolchain/
+cp install_java.sh sr-toolchain/
+
+copy_num=$(sed -n  '/===== Downloading thirdparty archives...done/=' sr-toolchain/starrocks/thirdparty/download-thirdparty.sh)
 if [[ copy_num == 0 ]]; then
     echo "===== cannot generate download scripts"
     exit 1
 fi
-head -n $copy_num starrocks/thirdparty/download-thirdparty.sh > starrocks/thirdparty/download-for-docker-thirdparty.sh
+head -n $copy_num sr-toolchain/starrocks/thirdparty/download-thirdparty.sh > sr-toolchain/starrocks/thirdparty/download-for-docker-thirdparty.sh
 
-# download thirdparty src
-if [[ ! -d "starrocks/thirdparty/src" ]]; then  
-    echo '========== download thirdparty src...'
-    bash starrocks/thirdparty/download-for-docker-thirdparty.sh
-fi 
+echo '========== download thirdparty src...'
+bash sr-toolchain/starrocks/thirdparty/download-for-docker-thirdparty.sh
 
-echo "========== start to build $IMAGE_NAME_TOOLCHAIN..."
-
-# build $IMAGE_NAME_TOOLCHAIN && copy thirdparty from CONTAINER(env_gen)
+cd sr-toolchain
 docker build \
 -t starrocks/$IMAGE_NAME_TOOLCHAIN:$IMAGE_VERSION \
 --build-arg GCC_VERSION=$GCC_VERSION \
 --build-arg GCC_URL=$GCC_URL \
---build-arg CMAKE_VERSION=$CMAKE_VERSION \
---build-arg CMAKE_DOWNLOAD_URL=$CMAKE_DOWNLOAD_URL \
 --build-arg MAVEN_VERSION=$MAVEN_VERSION \
 --build-arg SHA=$SHA \
 --build-arg BASE_URL=$BASE_URL .
@@ -85,33 +84,27 @@ echo "========== start to transfer thirdparty..."
 docker cp $CONTAINER_NAME_TOOLCHAIN:/var/local/thirdparty ../sr-thirdparty/
 rm -rf jdk.rpm
 
-cd ../sr-thirdparty
+cd ..
+
+cp java.tar.gz sr-thirdparty/
+if [[ ! -f "sr-thirdparty/java.tar.gz" ]]; then
+    echo "jdk not found"
+    exit 1
+fi
+
+cp -r cmake sr-thirdparty/
+cp install_env_gcc.sh sr-thirdparty/
+cp install_java.sh sr-thirdparty/
+
+# build thirdparty
+cd sr-thirdparty
 if [[ ! -d "thirdparty" ]]; then
     echo "thirdparty not found"
     exit 1
 fi
 rm -rf thirdparty/src
 
-cp ../jdk.rpm .
-if [[ ! -f "jdk.rpm" ]]; then
-    echo "jdk not found"
-    exit 1
-fi
-
-wget -O clang-llvm.tar.xz "$LLVM_SOURCE"
-
-if [[ ! -f "clang-llvm.tar.xz" ]]; then
-    echo "clang-llvm.tar.xz found"
-    exit 1
-fi
-
-# mkdir llvm && tar -xvJf clang-llvm.tar.xz -C llvm --strip-components 1
-
 mkdir -p llvm/bin && cp /home/disk1/doris-deps/toolchain/installed/llvm-10.0.1/bin/clang-format llvm/bin
-
-wget -O cmake.tar "$CMAKE_SOURCE"
-rm -rf cmake && mkdir cmake && tar -xvf cmake.tar -C cmake --strip-components 1
-rm -rf cmake.tar
 
 docker cp $CONTAINER_NAME_TOOLCHAIN:/usr/share/maven ../sr-thirdparty/
